add updrage script backend frontend
