+++ apps/frontend/store/slices/ChatStoreSlice.ts
@@ -55,6 +55,7 @@ export const sendMessage = createAsyncThunk<
     const userID = (getState() as RootState).user.userId;
 
     try {
+      if (!userID) return rejectWithValue("User not authenticated");
       const res = await api.post("/api/v1/query", { query: content, namespace: userID });
       const payload = res?.data;
       const answer =
