+++ apps/script/requirements.txt
@@ -22,3 +22,5 @@ functions-framework
 
 # System dependencies
 pysqlite3-binary
+transformers>=4.40.0
+torch>=2.2.0
